S1: split utils/db.py into a per-feature package
The 1417-LOC utils/db.py became utils/db/ — a package whose
submodules host per-feature CRUD.  utils/db/__init__.py re-exports
every public symbol so the historical import patterns

    from utils import db; db.func(...)
    from utils.db import func

continue to resolve without any caller-side change.

Layout
------
utils/db/
├── __init__.py    re-exports (pool, codec, every per-feature submodule)
├── pool.py        asyncpg pool lifecycle + generic CRUD primitives
├── codec.py       JSONB codec registration + legacy-decode shim
├── migrations.py  schema bootstrap + advisory-locked migration runner
├── xp.py          xp table CRUD + level math
├── settings.py    guild_settings KV
├── roles.py       role_thresholds + reaction_roles
├── moderation.py  warnings + mod_logs + prohibited_words
├── economy.py     coins + economy + job_progress
├── inventory.py   inventory CRUD
├── governance.py  subsystem_visibility + cleanup_policies + audit
├── sessions.py    runtime_sessions + runtime_session_state
├── anchors.py     panel_anchors
└── games/
    ├── __init__.py
    ├── rps.py
    ├── counting.py
    ├── chain.py
    ├── deathmatch.py
    └── mining.py
   (blackjack has no dedicated table — uses xp.coins via economy.py.)

Ownership rule preserved by structure
-------------------------------------
Submodules access primitives via ``from utils.db import pool`` and
call ``pool.get() / pool.fetchone() / pool.fetchall() / pool.execute()``.
This indirection means tests that monkeypatch ``utils.db.pool.<name>``
propagate to every submodule caller — backward-compatible with the
existing governance/test conftest pattern that patches names on the
``utils.db`` module namespace.

The legacy private alias ``_maybe_decode_legacy`` is preserved as a
re-export in sessions.py + __init__.py until S6 retires the shim
(blocked on production confirmation that migration 012 ran).

Behaviour preserved
-------------------
- All SQL strings preserved verbatim.
- All transaction boundaries preserved (deathmatch dual-update,
  set_session_state_many, set_mining_inventory).
- asyncpg JSONB codec registration unchanged.
- Migration advisory-lock behaviour unchanged.
- Public function signatures unchanged.

Tests
-----
New tests/unit/registry/test_db_package_surface.py:
- 87 public symbols × 2 import paths (``from utils import db; db.X``
  and ``from utils.db import X``) parameterised — 174 assertions.
- ``_maybe_decode_legacy`` private back-compat alias still reachable.
- Every submodule namespace imports cleanly.

Verification: ruff/black/isort/mypy all clean; pytest 419/419 green
(was 241; +178 from the new surface-regression tests).
